+++ packages/ag-grid-enterprise/src/serverSideRowModel/listeners/expandListener.ts
@@ -1,71 +1,26 @@
-import type { BeanCollection, NamedBean } from 'ag-grid-community';
-import { BeanStub, RowNode, _exists, _getRowHeightForNode, _isServerSideRowModel, _missing } from 'ag-grid-community';
+import type { NamedBean } from 'ag-grid-community';
+import { BeanStub, _isServerSideRowModel } from 'ag-grid-community';
 
-import type { ServerSideRowModel } from '../serverSideRowModel';
-import type { StoreFactory } from '../stores/storeFactory';
+import type { ServerSideExpansionService } from '../services/serverSideExpansionService';
 
 export class ExpandListener extends BeanStub implements NamedBean {
     beanName = 'ssrmExpandListener' as const;
 
-    private serverSideRowModel: ServerSideRowModel;
-    private storeFactory: StoreFactory;
-
-    public wireBeans(beans: BeanCollection) {
-        this.serverSideRowModel = beans.rowModel as ServerSideRowModel;
-        this.storeFactory = beans.ssrmStoreFactory as StoreFactory;
-    }
-
     public postConstruct(): void {
-        // only want to be active if SSRM active, otherwise would be interfering with other row models
         if (!_isServerSideRowModel(this.gos)) {
-            return;
+            return; // only want to be active if SSRM active, otherwise would be interfering with other row models
         }
 
         this.addManagedEventListeners({ rowExpansionStateChanged: this.onRowExpandStateChanged.bind(this) });
     }
 
     private onRowExpandStateChanged(): void {
-        this.beans.rowModel.forEachNode((rowNode) => {
-            if (rowNode.expanded) {
-                if (rowNode.master) {
-                    this.createDetailNode(rowNode);
-                }
-                if (_missing(rowNode.childStore) && rowNode.hasChildren()) {
-                    const storeParams = this.serverSideRowModel.getParams();
-                    rowNode.childStore = this.createBean(this.storeFactory.createStore(storeParams, rowNode));
-                }
-            } else if (this.gos.get('purgeClosedRowNodes') && _exists(rowNode.childStore)) {
-                rowNode.childStore = this.destroyBean(rowNode.childStore)!;
-            }
+        const beans = this.beans;
+        const expansionSvx = beans.expansionSvc as ServerSideExpansionService;
+        beans.rowModel.forEachNode((rowNode) => {
+            expansionSvx.updateExpandedState(rowNode);
         });
 
         this.eventSvc.dispatchEvent({ type: 'storeUpdated' });
-    }
-
-    private createDetailNode(masterNode: RowNode): RowNode {
-        if (_exists(masterNode.detailNode)) {
-            return masterNode.detailNode;
-        }
-
-        const detailNode = new RowNode(this.beans);
-
-        detailNode.detail = true;
-        detailNode.selectable = false;
-        detailNode.parent = masterNode;
-
-        if (_exists(masterNode.id)) {
-            detailNode.id = 'detail_' + masterNode.id;
-        }
-
-        detailNode.data = masterNode.data;
-        detailNode.level = masterNode.level + 1;
-
-        const defaultDetailRowHeight = 200;
-        const rowHeight = _getRowHeightForNode(this.beans, detailNode).height;
-
-        detailNode.rowHeight = rowHeight ? rowHeight : defaultDetailRowHeight;
-        masterNode.detailNode = detailNode;
-
-        return detailNode;
     }
 }
